Add util to turn LinearMaps into matrices
`convert` methods from `LinearMaps.jl` don't really work for us.
If we're just wrapping a matrix, we want to get that same matrix back.
This would require us to `convert` to `AbstractMatrix`. But then, if
it's not a wrapped map, it'll fall back to a `Matrix` conversion which
is horribly inefficient.
So now we have our own function for it.