Better, "soft" handling of boundary constraints
Rather than enforcing "hard" boundary constraints (which are fulfilled
exactly), the default is now that the constraints are fulfilled up to
noise of a certain, user-specified magnitude.

This enables us to keep the constrained DOFs in the GMRF without
reducing the rank of the precision matrix, and without resorting to
hacks like `ConstrainedGMRF`.

`ConstrainedGMRF` should be removed soon.