Poisson offset rework (#58)
* Move `binomial_observations.jl` to exp family dir

* Handle offsets via new `PoissonObservations` type

Previously, it was handled by passing a kwarg.
However the expected use case for observation models / likelihoods is
that the observations always stay the same and only the
hyperparameters (which are passed via kwargs) change between iterations.
So it's more intuitive to absorb the offset into the observations, since
the offset will also stay fixed in all applications I imagine.

* Add exposure option to formula interface